python/pyproject.toml
[tool.poetry]
name = "fixa"
version = "0.0.0"
description = "Python template"
authors = ["xrelkd <46590321+xrelkd@users.noreply.github.com>"]
readme = "README.md"
packages = [
  { include = "fixa", from = "src" },
  { include = "fixa2", from = "src" },
]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
flask    = "^3.0.0"
pillow   = "^10.1.0"
python   = "^3.10"
requests = "^2.31.0"

[tool.poetry.scripts]
fixa  = "fixa:main"
fixa2 = "fixa2:main"

[tool.ruff]
target-version = "py311"

line-length = 88

[tool.ruff.lint]
fixable = ["ALL"]
ignore = ["E501"]
select = [
  # pycodestyle
  "E",
  # pycodestyle
  "W",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # pep8-naming
  "N",
  # flake8-return
  "RET",
]

[tool.ruff.format]
quote-style               = "double"
indent-style              = "space"
skip-magic-trailing-comma = false
line-ending               = "auto"

[tool.black]
line-length = 88

[tool.isort]
line_length                    = 88   # to be consistent with yapf
multi_line_output              = 3
include_trailing_comma         = true
force_grid_wrap                = 0
use_parentheses                = true
ensure_newline_before_comments = true

[tool.mypy]
python_version         = "3.11"
pretty                 = true
warn_redundant_casts   = true
disallow_untyped_calls = true
disallow_untyped_defs  = true
no_implicit_optional   = true
